Adding start methods (see desc)
Added game life cycle for most sources.
This means the possibility to start, stop or kill a game from these sources :
- Retroarch
- Dolphin
- Citra
- Yuzu
- PPSSPP

Howerver some sources ensure that they keep only one instance and mess with life cycle handling, only start is implemented :
- Steam
- Legendary
- Lutris (default)

Lutris is fixable, with a direct call to the runners. For example, direclty call wine instead of calling through Lutris.
For now, Lutris games that use the wine runner are disabled for this reason, they will be implemented separately.