🎉 Starting conversion to python + gnome-builder
This is ✨ PAINFUL ✨
I've already lost multiple hours of work because I thought that using
gnome-builder AFTER writing python code would work out of the box.
Rookie mistake.

My motivation took a hit because of this.
Maybe a minimal viable product will be done this weekend.

The goal is simple, scanning Citra games.
Once this is implemented, the rest will flow.